+++ tests/testthat/test-utils_ask.R
@@ -9,9 +9,9 @@ tar_test("tar_ask_env()", {
 
 tar_test("tar_should_overwrite", {
   tmp <- tempfile()
-  expect_true(tar_should_overwrite(ask = TRUE, file = tmp))
-  expect_true(tar_should_overwrite(ask = FALSE, file = tmp))
+  expect_true(tar_should_overwrite(path = tmp, ask = TRUE))
+  expect_true(tar_should_overwrite(path = tmp, ask = FALSE))
   file.create(tmp)
-  expect_true(tar_should_overwrite(ask = FALSE, file = tmp))
+  expect_true(tar_should_overwrite(path = tmp, ask = FALSE))
   # More testing is in tests/interactive/test-tar_script.R # nolint.
 })